Stellar changes
Move part of code responsible for send income to asset holders from StellarServer.php to StellarOperator.php;
Extract code that fetches bets from Stellar and sends prizes to betters from StellarCroupierController.php to StellarCroupier.php;
Add docs to Croupier::prizeAmount
Fix StellarOperatorController.php: display $resultCode
