+++ commands/StellarCroupierController.php
@@ -5,9 +5,8 @@ namespace app\commands;
 use app\commands\traits\ControllerLogTrait;
 use app\interfaces\CronChainedInterface;
 use app\models\Croupier;
-use app\models\StellarServer;
+use app\models\StellarCroupier;
 use yii\console\Controller;
-use ZuluCrypto\StellarSdk\Model\Payment;
 
 /**
  * Class StellarCroupierController
@@ -19,8 +18,6 @@ use ZuluCrypto\StellarSdk\Model\Payment;
 class StellarCroupierController extends Controller implements CronChainedInterface
 {
     use ControllerLogTrait;
-
-    public const PRIZE_MEMO_TEXT = 'x%d Winner Prize';
 
     public function actionIndex()
     {
@@ -34,46 +31,32 @@ class StellarCroupierController extends Controller implements CronChainedInterfa
      */
     protected function sendGameProfits()
     {
-        $stellarServer = new StellarServer();
+        $stellarServer = new StellarCroupier();
 
         // TODO get sinceCursor
         $sinceCursor = null;
 
-        $croupierAccount = $stellarServer->getAccount($stellarServer->getCroupierPublicKey());
-
-        $payments = array_filter(
-            $croupierAccount->getPayments($sinceCursor), // gets at most 50 payments
-            fn ($p) =>
-                get_class($p) === Payment::class
-                && $p->getToAccountId() == StellarServer::getCroupierPublicKey()
-                && $p->isNativeAsset()
-        );
-
-        $croupierBalance = $croupierAccount->getNativeBalance();
         $winners_count = 0;
-        /** @var Payment $payment */
-        foreach ($payments as $payment) {
-            $playerPublicKey = $payment->getFromAccountId();
-            $betAmount = $payment->getAmount()->getBalance();
-            if ($result = Croupier::prizeAmount($betAmount, $croupierBalance)) {
+        $bets = $stellarServer->getBets($sinceCursor);
+        foreach ($bets as [
+                 'player_public_key' => $playerPublicKey,
+                 'bet_amount' => $betAmount,
+                 'paging_token' => $pagingToken,
+        ]) {
+            if ($result = Croupier::prizeAmount($betAmount, $stellarServer->getCroupierBalance())) {
                 if ($stellarServer->isTestnet()) {
                     $prizeAmount = 0.01;
                 } else {
                     $prizeAmount = $result['prize_amount'];
                 }
-                $response = $stellarServer
-                    ->buildTransaction(StellarServer::getCroupierPublicKey())
-                    ->addLumenPayment($playerPublicKey, $prizeAmount)
-                    ->setTextMemo(sprintf(self::PRIZE_MEMO_TEXT, $result['winner_rate']))
-                    ->submit(StellarServer::getCroupierPrivateKey());
-                $croupierBalance -= $prizeAmount + $response->getResult()->getFeeCharged()->getScaledValue();
 
+                $stellarServer->sendPrize($playerPublicKey, $prizeAmount, $result['winner_rate']);
                 $this->output('[Croupier] User with account ' . $playerPublicKey . 'has won XLM ' . number_format($prizeAmount, 6) . 'with rate x' . $result['winner_rate']);
                 $winners_count++;
             }
 
-            // TODO save $payment->getPagingToken()
+            // TODO save $pagingToken
         }
-        $this->output('[Croupier] ' . $winners_count . ' users won out of ' . count($payments));
+        $this->output('[Croupier] ' . $winners_count . ' users won out of ' . count($bets));
     }
 }
